Feature : Server call Linkages : Ensure year screen shows data from server with all data mapping done

+++ common/domainmodels/src/main/kotlin/com/demo/datausage/domainmodels/DataUsageModels.kt
@@ -5,13 +5,13 @@ enum class Datatype{
     MOBILE_DATA_USAGE
 }
 data class YearWiseData(
-    val year:Int,
+    val year:Long,
     val dataType:Datatype,
     val value:String
 )
 
 data class QuarterWiseData(
-    val year:Int,
+    val year:Long,
     val dataType:Datatype,
     val qOneValue:String,
     val qTwoValue:String,
